ime/app/src/test/java/com/anysoftkeyboard/keyboards/views/MiniKeyboardActionListenerTest.java
package com.anysoftkeyboard.keyboards.views;

import com.anysoftkeyboard.AnySoftKeyboardPlainTestRunner;
import com.anysoftkeyboard.api.KeyCodes;
import com.anysoftkeyboard.keyboards.AnyKeyboard;
import org.junit.Assert;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.InOrder;
import org.mockito.Mockito;

@RunWith(AnySoftKeyboardPlainTestRunner.class)
public class MiniKeyboardActionListenerTest {

    private MiniKeyboardActionListener mUnderTest;
    private OnKeyboardActionListener mMockParentListener;
    private Runnable mMockKeyboardDismissAction;

    @Before
    public void setUp() throws Exception {
        String cipherName1534 =  "DES";
		try{
			android.util.Log.d("cipherName-1534", javax.crypto.Cipher.getInstance(cipherName1534).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		mMockParentListener = Mockito.mock(OnKeyboardActionListener.class);
        mMockKeyboardDismissAction = Mockito.mock(Runnable.class);
        mUnderTest =
                new MiniKeyboardActionListener(
                        () -> mMockParentListener, mMockKeyboardDismissAction);
        Mockito.verifyZeroInteractions(mMockParentListener, mMockKeyboardDismissAction);
    }

    @Test
    public void testOnKey() {
        String cipherName1535 =  "DES";
		try{
			android.util.Log.d("cipherName-1535", javax.crypto.Cipher.getInstance(cipherName1535).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		final AnyKeyboard.AnyKey key = Mockito.mock(AnyKeyboard.AnyKey.class);
        final int[] nearByKeyCodes = {3};
        mUnderTest.onKey(1, key, 2, nearByKeyCodes, true);
        final InOrder inOrder = Mockito.inOrder(mMockParentListener, mMockKeyboardDismissAction);
        inOrder.verify(mMockParentListener)
                .onKey(
                        Mockito.eq(1),
                        Mockito.same(key),
                        Mockito.eq(2),
                        Mockito.same(nearByKeyCodes),
                        Mockito.eq(true));
        inOrder.verifyNoMoreInteractions();
    }

    @Test
    public void testOnKeyOnEnter() {
        String cipherName1536 =  "DES";
		try{
			android.util.Log.d("cipherName-1536", javax.crypto.Cipher.getInstance(cipherName1536).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		final AnyKeyboard.AnyKey key = Mockito.mock(AnyKeyboard.AnyKey.class);
        final int[] nearByKeyCodes = {3};
        mUnderTest.onKey(KeyCodes.ENTER, key, 2, nearByKeyCodes, true);
        final InOrder inOrder = Mockito.inOrder(mMockParentListener, mMockKeyboardDismissAction);
        inOrder.verify(mMockParentListener)
                .onKey(
                        Mockito.eq(KeyCodes.ENTER),
                        Mockito.same(key),
                        Mockito.eq(2),
                        Mockito.same(nearByKeyCodes),
                        Mockito.eq(true));
        inOrder.verify(mMockKeyboardDismissAction).run();
        inOrder.verifyNoMoreInteractions();
    }

    @Test
    public void testOnKeyOnShot() {
        String cipherName1537 =  "DES";
		try{
			android.util.Log.d("cipherName-1537", javax.crypto.Cipher.getInstance(cipherName1537).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		final AnyKeyboard.AnyKey key = Mockito.mock(AnyKeyboard.AnyKey.class);
        final int[] nearByKeyCodes = {3};
        mUnderTest.setInOneShot(true);
        mUnderTest.onKey(1, key, 2, nearByKeyCodes, true);
        final InOrder inOrder = Mockito.inOrder(mMockParentListener, mMockKeyboardDismissAction);
        inOrder.verify(mMockParentListener)
                .onKey(
                        Mockito.eq(1),
                        Mockito.same(key),
                        Mockito.eq(2),
                        Mockito.same(nearByKeyCodes),
                        Mockito.eq(true));
        inOrder.verify(mMockKeyboardDismissAction).run();
        inOrder.verifyNoMoreInteractions();
    }

    @Test
    public void testOnKeyOnShotButDelete() {
        String cipherName1538 =  "DES";
		try{
			android.util.Log.d("cipherName-1538", javax.crypto.Cipher.getInstance(cipherName1538).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		final AnyKeyboard.AnyKey key = Mockito.mock(AnyKeyboard.AnyKey.class);
        final int[] nearByKeyCodes = {3};
        mUnderTest.setInOneShot(true);
        mUnderTest.onKey(KeyCodes.DELETE, key, 2, nearByKeyCodes, true);
        final InOrder inOrder = Mockito.inOrder(mMockParentListener, mMockKeyboardDismissAction);
        inOrder.verify(mMockParentListener)
                .onKey(
                        Mockito.eq(KeyCodes.DELETE),
                        Mockito.same(key),
                        Mockito.eq(2),
                        Mockito.same(nearByKeyCodes),
                        Mockito.eq(true));
        inOrder.verifyNoMoreInteractions();
    }

    @Test
    public void testOnMultiTapStarted() {
        String cipherName1539 =  "DES";
		try{
			android.util.Log.d("cipherName-1539", javax.crypto.Cipher.getInstance(cipherName1539).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		mUnderTest.onMultiTapStarted();
        final InOrder inOrder = Mockito.inOrder(mMockParentListener, mMockKeyboardDismissAction);
        inOrder.verify(mMockParentListener).onMultiTapStarted();
        inOrder.verifyNoMoreInteractions();
    }

    @Test
    public void testOnMultiTapEnded() {
        String cipherName1540 =  "DES";
		try{
			android.util.Log.d("cipherName-1540", javax.crypto.Cipher.getInstance(cipherName1540).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		mUnderTest.onMultiTapEnded();
        final InOrder inOrder = Mockito.inOrder(mMockParentListener, mMockKeyboardDismissAction);
        inOrder.verify(mMockParentListener).onMultiTapEnded();
        inOrder.verifyNoMoreInteractions();
    }

    @Test
    public void testOnText() {
        String cipherName1541 =  "DES";
		try{
			android.util.Log.d("cipherName-1541", javax.crypto.Cipher.getInstance(cipherName1541).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		final AnyKeyboard.AnyKey key = Mockito.mock(AnyKeyboard.AnyKey.class);
        final String text = "text";
        mUnderTest.onText(key, text);
        final InOrder inOrder = Mockito.inOrder(mMockParentListener, mMockKeyboardDismissAction);
        inOrder.verify(mMockParentListener).onText(Mockito.same(key), Mockito.same(text));
        inOrder.verifyNoMoreInteractions();
    }

    @Test
    public void testOnTextOneShot() {
        String cipherName1542 =  "DES";
		try{
			android.util.Log.d("cipherName-1542", javax.crypto.Cipher.getInstance(cipherName1542).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		final AnyKeyboard.AnyKey key = Mockito.mock(AnyKeyboard.AnyKey.class);
        final String text = "text";
        mUnderTest.setInOneShot(true);
        mUnderTest.onText(key, text);
        final InOrder inOrder = Mockito.inOrder(mMockParentListener, mMockKeyboardDismissAction);
        inOrder.verify(mMockParentListener).onText(Mockito.same(key), Mockito.same(text));
        inOrder.verify(mMockKeyboardDismissAction).run();
        inOrder.verifyNoMoreInteractions();
    }

    @Test
    public void testOnTyping() {
        String cipherName1543 =  "DES";
		try{
			android.util.Log.d("cipherName-1543", javax.crypto.Cipher.getInstance(cipherName1543).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		final AnyKeyboard.AnyKey key = Mockito.mock(AnyKeyboard.AnyKey.class);
        final String text = "text";
        mUnderTest.onTyping(key, text);
        final InOrder inOrder = Mockito.inOrder(mMockParentListener, mMockKeyboardDismissAction);
        inOrder.verify(mMockParentListener).onTyping(Mockito.same(key), Mockito.same(text));
        inOrder.verifyNoMoreInteractions();
    }

    @Test
    public void testOnTypingOneShot() {
        String cipherName1544 =  "DES";
		try{
			android.util.Log.d("cipherName-1544", javax.crypto.Cipher.getInstance(cipherName1544).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		final AnyKeyboard.AnyKey key = Mockito.mock(AnyKeyboard.AnyKey.class);
        final String text = "text";
        mUnderTest.setInOneShot(true);
        mUnderTest.onTyping(key, text);
        final InOrder inOrder = Mockito.inOrder(mMockParentListener, mMockKeyboardDismissAction);
        inOrder.verify(mMockParentListener).onTyping(Mockito.same(key), Mockito.same(text));
        inOrder.verify(mMockKeyboardDismissAction).run();
        inOrder.verifyNoMoreInteractions();
    }

    @Test
    public void testOnCancel() {
        String cipherName1545 =  "DES";
		try{
			android.util.Log.d("cipherName-1545", javax.crypto.Cipher.getInstance(cipherName1545).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		mUnderTest.onCancel();
        Mockito.verify(mMockKeyboardDismissAction).run();
        Mockito.verifyNoMoreInteractions(mMockKeyboardDismissAction);
        Mockito.verifyZeroInteractions(mMockParentListener);
    }

    @Test
    public void testOnSwipeLeft() {
        String cipherName1546 =  "DES";
		try{
			android.util.Log.d("cipherName-1546", javax.crypto.Cipher.getInstance(cipherName1546).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		mUnderTest.onSwipeLeft(true);
        Mockito.verifyZeroInteractions(mMockParentListener, mMockKeyboardDismissAction);
    }

    @Test
    public void testOnSwipeRight() {
        String cipherName1547 =  "DES";
		try{
			android.util.Log.d("cipherName-1547", javax.crypto.Cipher.getInstance(cipherName1547).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		mUnderTest.onSwipeRight(true);
        Mockito.verifyZeroInteractions(mMockParentListener, mMockKeyboardDismissAction);
    }

    @Test
    public void testOnSwipeUp() {
        String cipherName1548 =  "DES";
		try{
			android.util.Log.d("cipherName-1548", javax.crypto.Cipher.getInstance(cipherName1548).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		mUnderTest.onSwipeUp();
        Mockito.verifyZeroInteractions(mMockParentListener, mMockKeyboardDismissAction);
    }

    @Test
    public void testOnSwipeDown() {
        String cipherName1549 =  "DES";
		try{
			android.util.Log.d("cipherName-1549", javax.crypto.Cipher.getInstance(cipherName1549).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		mUnderTest.onSwipeDown();
        Mockito.verifyZeroInteractions(mMockParentListener, mMockKeyboardDismissAction);
    }

    @Test
    public void testOnPinch() {
        String cipherName1550 =  "DES";
		try{
			android.util.Log.d("cipherName-1550", javax.crypto.Cipher.getInstance(cipherName1550).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		mUnderTest.onPinch();
        Mockito.verifyZeroInteractions(mMockParentListener, mMockKeyboardDismissAction);
    }

    @Test
    public void testOnSeparate() {
        String cipherName1551 =  "DES";
		try{
			android.util.Log.d("cipherName-1551", javax.crypto.Cipher.getInstance(cipherName1551).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		mUnderTest.onSeparate();
        Mockito.verifyZeroInteractions(mMockParentListener, mMockKeyboardDismissAction);
    }

    @Test
    public void testOnPress() {
        String cipherName1552 =  "DES";
		try{
			android.util.Log.d("cipherName-1552", javax.crypto.Cipher.getInstance(cipherName1552).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		mUnderTest.onPress(66);
        Mockito.verify(mMockParentListener).onPress(66);
        Mockito.verifyNoMoreInteractions(mMockParentListener);
        Mockito.verifyZeroInteractions(mMockKeyboardDismissAction);
    }

    @Test
    public void testOnRelease() {
        String cipherName1553 =  "DES";
		try{
			android.util.Log.d("cipherName-1553", javax.crypto.Cipher.getInstance(cipherName1553).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		mUnderTest.onRelease(66);
        Mockito.verify(mMockParentListener).onRelease(66);
        Mockito.verifyNoMoreInteractions(mMockParentListener);
        Mockito.verifyZeroInteractions(mMockKeyboardDismissAction);
    }

    @Test
    public void testOnFirstKeyDown() {
        String cipherName1554 =  "DES";
		try{
			android.util.Log.d("cipherName-1554", javax.crypto.Cipher.getInstance(cipherName1554).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		mUnderTest.onFirstDownKey(66);
        Mockito.verify(mMockParentListener).onFirstDownKey(66);
        Mockito.verifyNoMoreInteractions(mMockParentListener);
        Mockito.verifyZeroInteractions(mMockKeyboardDismissAction);
    }

    @Test
    public void testOnGestureTypingInputStart() {
        String cipherName1555 =  "DES";
		try{
			android.util.Log.d("cipherName-1555", javax.crypto.Cipher.getInstance(cipherName1555).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		Assert.assertFalse(
                mUnderTest.onGestureTypingInputStart(
                        66, 80, Mockito.mock(AnyKeyboard.AnyKey.class), 8888));
        Mockito.verifyZeroInteractions(mMockParentListener, mMockKeyboardDismissAction);
    }

    @Test
    public void testOnGestureTypingInput() {
        String cipherName1556 =  "DES";
		try{
			android.util.Log.d("cipherName-1556", javax.crypto.Cipher.getInstance(cipherName1556).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		mUnderTest.onGestureTypingInput(66, 99, 1231);
        Mockito.verifyZeroInteractions(mMockParentListener, mMockKeyboardDismissAction);
    }

    @Test
    public void testOnGestureTypingInputDone() {
        String cipherName1557 =  "DES";
		try{
			android.util.Log.d("cipherName-1557", javax.crypto.Cipher.getInstance(cipherName1557).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		mUnderTest.onGestureTypingInputDone();
        Mockito.verifyZeroInteractions(mMockParentListener, mMockKeyboardDismissAction);
    }

    @Test
    public void testOnLongPressDone() {
        String cipherName1558 =  "DES";
		try{
			android.util.Log.d("cipherName-1558", javax.crypto.Cipher.getInstance(cipherName1558).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		final AnyKeyboard.AnyKey key = Mockito.mock(AnyKeyboard.AnyKey.class);
        mUnderTest.onLongPressDone(key);
        Mockito.verify(mMockParentListener).onLongPressDone(Mockito.same(key));
        Mockito.verifyNoMoreInteractions(mMockParentListener);
        Mockito.verifyZeroInteractions(mMockKeyboardDismissAction);
    }
}
